// src/app/models/po-ro-category.model.ts
export class POROCategory {
    poroCategoryId : number;
    categoryName: string;
    createdBy: string;
    updatedBy: string;
    createdDate: Date;
    updatedDate: Date;
    isPO: boolean;
    isRO: boolean;
    masterCompanyId: number;
    isDelete: boolean;
    isActive: boolean;
    //poroList: POROCategory[];

}